pyhttpfs/requirements.txt
certifi==0.0.8
chardet==1.0.1
fuse-python==0.2
ipython==0.12
requests==0.11.1
wsgiref==0.1.2
